+++ src/workspace/src/led_controller/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.0.2)
-project(led_hydroball)
+project(led_controller)
 
 ## Compile as C++11, supported in ROS Kinetic and newer
 # add_compile_options(-std=c++11)
@@ -28,7 +28,7 @@ find_package(catkin REQUIRED COMPONENTS
 ## DEPENDS: system dependencies of this project that dependent projects also need
 catkin_package(
   INCLUDE_DIRS include
-  LIBRARIES led_hydroball
+  LIBRARIES led_controller
   CATKIN_DEPENDS roscpp std_msgs led_service
   DEPENDS system_lib
 )
@@ -46,7 +46,7 @@ include_directories(
 
 # Declare a C++ library
  add_library(${PROJECT_NAME}
-   src/hydroball_led.cpp
+   src/led_controller.cpp
  )
 
 ## Add cmake target dependencies of the library
@@ -57,7 +57,7 @@ include_directories(
 ## Declare a C++ executable
 ## With catkin_make all packages are built within a single CMake context
 # The recommended prefix ensures that target names across packages don't collide
- add_executable(${PROJECT_NAME}_node node/hydroball_led_node.cpp)
+ add_executable(${PROJECT_NAME}_node node/led_controller_node.cpp)
 
 ## Rename C++ executable without prefix
 ## The above recommended prefix causes long target names, the following renames the
