+++ src/workspace/src/led_controller/node/led_controller_node.cpp
@@ -1,5 +1,5 @@
 #include "ros/ros.h"
-#include "led_hydroball/hydroball_led.h"
+#include "led_controller/led_controller.h"
 
 int main(int argc, char **argv)
 {
