+++ src/workspace/src/led_controller/package.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>led_hydroball</name>
+  <name>led_controller</name>
   <version>0.0.0</version>
-  <description>The led_hydroball package</description>
+  <description>The led_controller package</description>
 
   <!-- One maintainer tag required, multiple allowed, one person per tag -->
   <!-- Example:  -->
