--- src/workspace/src/led_hydroball/src/hydroball_led.cpp
@@ -1,1 +1,0 @@
-#include "led_hydroball/hydroball_led.h"
